Task3/requirements.txt
torch
torchvision
transformers
google-generativeai
scikit-learn
matplotlib
gradio
